% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_access.R
\name{add_access}
\alias{add_access}
\title{Add Bioconductor_User group to workspace access}
\usage{
add_access(namespace, name, dry.run = FALSE)
}
\arguments{
\item{namespace}{character(1) namespace (billing account) under
which the workspace belongs.}

\item{name}{character(1) name of the workspace to add access
credentials.}

\item{dry.run}{\code{logical(1)} When \code{TRUE}, no changes are made to
the workspace ACL. A message is printed describing what would
have been done.}
}
\value{
\code{add_access()} returns TRUE, invisibly.
}
\description{
\code{add_access()} adds the
\code{Bioconductor_User} group to a workspace with \code{READER}
permissions. Users gain access to the workspace (and others) by
being added to the Bioconductor_User group.
}
\examples{
add_access("landmarkanvil2", "Bioconductor-Package-AnVILHCAR", TRUE)
}
