% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILPublish-utils.R
\name{create_workspace}
\alias{create_workspace}
\title{Create an AnVIL Workspace}
\usage{
create_workspace(namespace, name)
}
\arguments{
\item{namespace}{\code{character(1)} AnVIL namespace (billing project)
to be used.}

\item{name}{\code{character(1)} AnVIL workspace name.}
}
\value{
\code{create_workspace()} returns \code{TRUE} invisibly if the workspace was
created successfully; otherwise, an error is raised.
}
\description{
Call the \code{Terra()} API to create a new AnVIL workspace. This is
a helper function for \code{as_workspace()}, but can be used directly if you
want to create a workspace without populating it with content from an R
package. This is typically used by developers who want to create a
workspace and then run other functions to populate it with content, e.g.,
importing data into the workspace.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
create_workspace("my-namespace", "my-workspace")
\dontshow{\}) # examplesIf}
}
