% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workspace.R
\name{as_workspace}
\alias{as_workspace}
\title{Render R packages as AnVIL workspaces}
\usage{
as_workspace(
  path,
  namespace,
  name = NULL,
  create = FALSE,
  update = FALSE,
  use_readme = FALSE,
  type = c("ipynb", "rmd", "both")
)
}
\arguments{
\item{path}{\code{character(1)} path to the location of the package
source code.}

\item{namespace}{\code{character(1)} AnVIL namespace (billing project)
to be used.}

\item{name}{\code{character(1)} AnVIL workspace name or NULL. If NULL,
the workspace name is set to
\code{"Bioconductor-Package-<pkgname>"}, where \verb{<pkgname>} is the
name of the package (from the DESCRIPTION file) at \code{path}.}

\item{create}{\code{logical(1)} Create a new project?}

\item{update}{\code{logical(1)} Update (over-write the existing
DASHBOARD and any similarly named notebooks) an existing
workspace?  If niether \code{create} nore \code{update} is TRUE, the code
to create a workspace is run but no output generated; this can
be useful during debugging.}

\item{use_readme}{\code{logical(1)} Defaults to \code{FALSE}; if \code{TRUE} the
content of README.md in package top-level folder is used with
the package \code{DESCRIPTION} version and provenance metadata for
rendering in the workspace 'DASHBOARD'.}

\item{type}{\code{character(1)} The type of notebook to be in the
workspace. Must be one of \code{ipynb}, \code{rmd}, or \code{both}.}
}
\value{
\code{as_workspace()} returns the URL of the updated workspace,
invisibly.
}
\description{
\code{as_workspace()} renders a package source tree (e.g.,
from a git checkout) as an AnVIL workspace.
}
\details{
Information from the DESCRIPTION file and Rmd YAML files
are used to populate the 'DASHBOARD' tab.\preformatted{See `?as_notebook()` for details on how vignettes are processed
to notebooks, and the limitations of the current approach.
}
}
