% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_as_workspace.R
\name{package_source_as_workspace}
\alias{package_source_as_workspace}
\title{Render R packages as AnVIL workspaces}
\usage{
package_source_as_workspace(
  path,
  namespace,
  name,
  create = FALSE,
  update = FALSE
)
}
\arguments{
\item{path}{\code{character(1)} path to the location of the package
source code.}

\item{namespace}{\code{character(1)} AnVIL namespace (billing project)
to be used.}

\item{name}{\code{character(1)} AnVIL workspace name.}

\item{create}{\code{logical(1)} Create a new project?}

\item{update}{\code{logical(1)} Update (over-write teh existing
DASHBOARD and any similarly named notebooks) an existing
workspace?  One of \code{create} and \code{update} must be TRUE.}
}
\value{
\code{package_source_as_workspace()} returns the URL of the
updated workspace, invisibly.
}
\description{
\code{package_source_as_workspace()} renders a package
source tree (e.g., from a git checkout) as an AnVIL workspace.
}
\details{
Information from the DESCRIPTION file and Rmd YAML files
are used to populate the 'DASHBOARD' tab.\preformatted{See `?vignettes_to_notebooks()` for details on how vignettes
are processed to notebooks, and the limitations of the current
approach.
}
}
