% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes_to_notebooks.R
\name{vignettes_to_notebooks}
\alias{vignettes_to_notebooks}
\title{Render vignettes as .ipynb notebooks}
\usage{
vignettes_to_notebooks(path, namespace, name, create = FALSE, update = FALSE)
}
\arguments{
\item{path}{\code{character(1)} path to the location of the package
source code.}

\item{namespace}{\code{character(1)} AnVIL namespace (billing project)
to be used.}

\item{name}{\code{character(1)} AnVIL workspace name.}

\item{create}{\code{logical(1)} Create a new project?}

\item{update}{\code{logical(1)} Update (over-write any similarly named
notebooks) an existing workspace?  One of \code{create} and \code{update}
must be TRUE.}
}
\description{
\code{vignettes_to_notebooks()} renders .Rmd vignettes as
.ipynb notebooks, and updates the notebooks in an AnVIL
workspace.
}
\details{
\code{.Rmd} Vignettes are processed to \code{.md} using
\code{rmarkdown::render(..., md_document())}, and then translated to
\code{.ipynb} using python software called \code{notedown}; notedown is
available at https://github.com/aaren/notedown.\preformatted{The translation is not perfect, for instance code chunks marked
as `eval = FALSE` are not marked as such in the python notebook.
}
}
